refact(models): define uint64 for ids and use Address model on User

+++ internal/models/Application.go
@@ -11,9 +11,9 @@ const (
 )
 
 type Application struct {
-	ID          uint       `json:"id"`
-	JobId       uint       `json:"job_id"`
-	CandidateId uint       `json:"candidate_id"`
+	ID          uint64     `json:"id"`
+	JobId       uint64     `json:"job_id"`
+	CandidateId uint64     `json:"candidate_id"`
 	AppliedAt   time.Time  `json:"applied_at"`
 	Status      StatusType `json:"status"`
 }
